refactor: remove `TempProvider`
